Refactor FunctionalSpec
